backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
cohere==4.37.0
supabase==2.3.0
requests==2.31.0
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy<2.0.0
httpx>=0.24.0,<0.25.0
pydantic==2.5.0
regex==2023.10.3
